feat: Make it possible to disable cleanup

+++ ci-scripts/load-test.sh
@@ -16,4 +16,6 @@ info "Benchmark"
 time ./benchmark-tekton.sh --total "${TEST_TOTAL:-100}" --concurrent "${TEST_CONCURRENT:-10}" --run "${TEST_RUN:-./run.yaml}" --debug
 
 info "Cleanup"
-kubectl delete --all PipelineRuns
+if ${TEST_DO_CLEANUP:-true}; then
+    kubectl delete --all PipelineRuns
+fi
